requirements.txt
fastapi==0.104.0
uvicorn==0.23.1
python-multipart==0.0.6
requests==2.31.0

# RAG stack (Python 3.11 compatible)
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.0
huggingface-hub==0.14.1
transformers==4.30.2

groq
